Rename metadata_parsing
